Re-Designed gatasconnect-udp to see if this one want's to go therough the Android Hotspot
