+++ src/lib/bluetooth/ace/src/bluetooth.cpp
@@ -148,7 +148,7 @@ void Bluetooth::on_receive_unknown(const etl::imessage &msg)
 void Bluetooth::on_receive(const GATAS::OwnshipPositionMsg &msg)
 {
     (void)msg;
-    // ownshipPosition = SpinlockGuard::withLock(spinLock, msg.position.assignTo());
+    // ownshipPosition = SpinlockGuard::withLock(spinLock, msg.position).assignTo();
 }
 
 void Bluetooth::createAdvData()
