+++ src/lib/dataport/ace/src/dataport.cpp
@@ -18,7 +18,7 @@ void DataPort::on_receive(const GATAS::OwnshipPositionMsg &msg)
     (void)msg;
     static Every<uint32_t, 500'000, 1'000'000> sendValidGps{0};
 
-    //    ownshipPosition = SpinlockGuard::withLock(spinLock, msg.position.assignTo());
+    //    ownshipPosition = SpinlockGuard::withLock(spinLock, msg.position).assignTo();
     // if (sendValidGps.isItTime(CoreUtils::timeUs32Raw()))
     // {
