+++ src/lib/gatasconnect/CMakeLists.txt
@@ -4,7 +4,7 @@ project(gatasconnect VERSION 0.0.0 LANGUAGES CXX)
 set(MSG_PREFIX "${PROJECT_NAME} |")
 
 set(MODULE_SOURCE_FILES
-    ace/src/gatasconnecttcp.cpp
+        ace/src/gatasconnectudp.cpp
 )
 
 set(MODULE_TARGET_LINK
